remove nanoaod (old NanoEvents) package, factor nanoevents for further expansion
