+++ coffea/nanoevents/mapping/__init__.py
@@ -1,0 +1,7 @@
+from .uproot import TrivialUprootOpener, UprootSourceMapping
+from .parquet import TrivialParquetOpener, ParquetSourceMapping
+from .cached import CachedMapping
+
+__all__ = ['TrivialUprootOpener', 'UprootSourceMapping',
+           'TrivialParquetOpener', 'ParquetSourceMapping',
+           'CachedMapping']
