+++ coffea/nanoevents/schemas/__init__.py
@@ -1,0 +1,5 @@
+from .base import BaseSchema
+from .nanoaod import NanoAODSchema
+from .treemaker import TreeMakerSchema
+
+__all__ = ['BaseSchema', 'NanoAODSchema', 'TreeMakerSchema']
