+++ coffea/processor/test_items/__init__.py
@@ -1,7 +1,6 @@
 from .NanoTestProcessor import NanoTestProcessor
-from .NanoEvents0Processor import NanoEvents0Processor
 from .NanoEventsProcessor import NanoEventsProcessor
 
 from .NanoTestProcessorPandas import NanoTestProcessorPandas
 
-__all__ = ["NanoTestProcessor", "NanoEvents0Processor", "NanoEventsProcessor", "NanoTestProcessorPandas"]
+__all__ = ["NanoTestProcessor", "NanoEventsProcessor", "NanoTestProcessorPandas"]
